mcount: Use 32-bit PLT index in the mcount rstack
For a large program, it might be possible to have more than 64K of PLT
calls.  In that case deciding PLT call by the index will be broken.  In
the other hand, we don't support filter depth more than 64K, just
exchange the type of filter depth field and dyn_idx.

+++ libmcount/internal.h
@@ -34,8 +34,8 @@ enum filter_result {
 struct filter_control {
 	int in_count;
 	int out_count;
-	int depth;
-	int saved_depth;
+	uint16_t depth;
+	uint16_t saved_depth;
 	uint64_t time;
 	uint64_t saved_time;
 };
